Allow subpixel movement for legacy effects
This won't change the movement speed, as the total movement
is recalculated every frame, so this shouldn't negatively affect
VSFilter compatibility; if anything this should slightly improve
compatibility since VSFilter works in a supersampled space.
In any case, it does improves visual quality.

+++ libass/ass_render.h
@@ -261,7 +261,7 @@ typedef struct {
         SCROLL_TB,              // top-to-bottom
         SCROLL_BT
     } scroll_direction;         // for EVENT_HSCROLL, EVENT_VSCROLL
-    int scroll_shift;
+    double scroll_shift;
     int scroll_y0, scroll_y1;
 
     // face properties
